backend/pyproject.toml
[project]
name = "rag-assistant-backend"
version = "0.2.1"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "email-validator>=2.0.0",
  "starlette>=0.40.0",
  "google-genai>=1.0.0",
  "google-cloud-storage>=2.18.0",
  "orjson>=3.10.7",
  "sqlalchemy>=2.0.32",
  "alembic>=1.13.2",
  "psycopg2-binary>=2.9.9",
  "greenlet",
  "prometheus-client>=0.20.0",
  "httpx>=0.27.0",
  "tenacity>=9.0.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "mypy>=1.11.0",
  "ruff>=0.6.8",
  "redis>=5.0.0",
  "arq>=0.26.0",
  "pyyaml>=6.0.0",
  "watchdog>=3.0.0",
]

[project.optional-dependencies]
test = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.1.0",
  "httpx>=0.24.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*", "scripts*"]

[tool.uvicorn]
factory = true
host = "0.0.0.0"
port = 8000

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_equality = true
